requirements.txt
asgiref==3.8.1
awscli==1.36.28
boto3==1.35.87
botocore==1.35.87
colorama==0.4.6
Django==5.0.9
django-storages==1.14.4
docutils==0.16
jmespath==1.0.1
pillow==10.4.0
psycopg==3.1.18
pyasn1==0.6.1
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
rsa==4.7.2
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.3.0
